refactor: create a separate show more shows button for the main page
