+++ src/components/pages/MainPage/MainPage.module.scss
@@ -6,11 +6,3 @@
   align-items: center;
   padding: 40px 0px;
 }
-
-.more {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 40px;
-  width: 100%;
-}
