+++ src/components/pages/MainPage/ShowMoreShowsButton/ShowMoreShowsButton.module.scss
@@ -1,0 +1,9 @@
+@import '../../../../styles/variables.scss';
+
+.more {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 40px;
+  width: 100%;
+}
